Added Studios page/ Added initial booking functionality/ Added images/ Added availability

+++ database/migrations/2024_11_26_162609_create_classes_table.php
@@ -11,21 +11,24 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('classes', function (Blueprint $table) {
-            $table->id();
-            $table->timestamps();
-            $table->string('name');
-            $table->string('genre');
-            $table->text('description')->nullable();
-            $table->dateTime('time_start');
-            $table->dateTime('time_ends');
-            $table->decimal('price', 8, 2);
-            $table->unsignedBigInteger('studios_id');
-            $table->unsignedBigInteger('instructors_id');
-
-            $table->foreign('studios_id')->references('id')->on('studios')->onDelete('cascade');
-            $table->foreign('instructors_id')->references('id')->on('instructors')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('classes')) {
+            Schema::create('classes', function (Blueprint $table) {
+                $table->id();
+                $table->timestamps();
+                $table->string('name');
+                $table->string('genre');
+                $table->text('description')->nullable();
+                $table->dateTime('time_start');
+                $table->dateTime('time_ends');
+                $table->decimal('price', 8, 2);
+                $table->unsignedBigInteger('studios_id');
+                $table->unsignedBigInteger('instructors_id');
+    
+                // Add foreign keys if needed
+                $table->foreign('studios_id')->references('id')->on('studios')->onDelete('cascade');
+                $table->foreign('instructors_id')->references('id')->on('instructors')->onDelete('cascade');
+            });
+        }
     }
 
     /**
